+++ src/App.css
@@ -25,6 +25,22 @@ h1 {
 nav, footer {
   background-color: rgba(0, 0, 0, 0.5);
   width: 100%;
+  display: flex;
+  justify-content: space-between;
+}
+
+.nav-title {
+  margin-left: 5vw;
+}
+
+nav a {
+  font-size: 20px;
+  margin-right: 5vw;
+  font-family: Arial, Helvetica, sans-serif;
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
 }
 
 p {
@@ -155,6 +171,7 @@ a {
 
 a:hover {
   color: #3e8e41;
+  cursor: pointer;
 }
 
 .about-text {
